requirements.txt
alembic==1.9.2
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
distlib==0.3.6
docker==6.0.1
exceptiongroup==1.1.0
fastapi==0.89.1
filelock==3.9.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
identify==2.5.16
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.1
pycparser==2.21
pydantic==1.10.4
pyinstrument==4.4.0
pytest==7.2.1
python-dotenv==0.21.0
PyYAML==6.0
redis==4.4.2
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.39.0
sqlalchemy2-stubs==0.0.2a31
sqlmodel==0.0.8
sqlparse==0.4.3
starlette==0.22.0
tomli==2.0.1
types-pyOpenSSL==23.0.0.2
types-redis==4.4.0.4
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.17.1
watchfiles==0.18.1
websocket-client==1.4.2
websockets==10.4
